% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate percent presence of genes}
\usage{
calc_stats(df)
}
\arguments{
\item{df}{The data frame holding the processed ARIBA results in long format}
}
\description{
Function that calculates the percent presence of all genes in the input report
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
