% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flags.R
\name{check_flags}
\alias{check_flags}
\title{ARIBA Flag evaluation function}
\usage{
check_flags(df, acquired = TRUE)
}
\arguments{
\item{df}{The data frame holding the flag values}

\item{acquired}{if the data is from acquired resistance screening, set this to TRUE, as different flags are accepted for the acquired genes compared to mutations in intrinsic genes.}
}
\description{
Function that evaluates the flags reported by ARIBA as "1", accepted, or "0", not accepted.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
