% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a presence/absence report}
\usage{
create_report(df, mut = FALSE)
}
\arguments{
\item{df}{The data frame holding the processed ARIBA results in long format}
}
\description{
Function that creates a presence, 1, absence, 0, report, one row per sample, one column per gene
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
