% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Generate overview table of genes}
\usage{
create_table(df, acquired = TRUE)
}
\arguments{
\item{df}{The data frame holding the raw ARIBA data}

\item{acquired}{logical, if the data is from acquired gene analysis then this should be set to TRUE}
}
\description{
Function that filters raw ARIBA data into a table with presence/absence results
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
