% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{filter_table}
\alias{filter_table}
\title{Filter table based on a list of genes}
\usage{
filter_table(df, list_genes)
}
\arguments{
\item{df}{The data frame}

\item{list_genes}{A vector of genes of interest}
}
\description{
Function that filters out all genes except the ones specified in list_genes supplied to the function
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
