% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gene_names.R
\name{fix_gene_names}
\alias{fix_gene_names}
\title{Correct gene names}
\usage{
fix_gene_names(df, ending, db)
}
\arguments{
\item{df}{The ARIBA result data}

\item{ending}{The file suffix of the input files from ARIBA}
}
\description{
Function that extracts the gene name from the column "cluster" in the ARIBA results.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
