% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gyr_par_results.R
\name{fix_gyr_par_results}
\alias{fix_gyr_par_results}
\title{Filter out non-QRDR mutations}
\usage{
fix_gyr_par_results(df)
}
\arguments{
\item{df}{The data frame holding the processed ARIBA results in long format}
}
\description{
Function that filters out all non-QRDR mutations detected in the dataset, only for gyrA, gyrB, parC, and parE
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
