% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_presence.R
\name{percent_presence}
\alias{percent_presence}
\title{Calculate the percent presence of a value}
\usage{
percent_presence(df, group_col)
}
\description{
Function that calculates the percent presence of all unique values in a column in relation to the total number of rows in the data frame (samples)
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
